requirements.txt
pandas
openpyxl
pywhatkit
phonenumbers
python-dotenv
reportlab
pyautogui
pygetwindow
pyinstaller